+++ src/main/java/com/example/c_design/req/GoodsQueryReq.java
@@ -1,7 +1,7 @@
 package com.example.c_design.req;
 
 public class GoodsQueryReq {
-    private Integer gId;
+    private Long gId;
 
     private String gName;
 
@@ -11,11 +11,11 @@ public class GoodsQueryReq {
 
     private Integer gNumber;
 
-    public Integer getgId() {
+    public Long getgId() {
         return gId;
     }
 
-    public void setgId(Integer gId) {
+    public void setgId(Long gId) {
         this.gId = gId;
     }
