+++ src/main/java/com/example/c_design/resp/StoreQueryResp.java
@@ -1,7 +1,7 @@
 package com.example.c_design.resp;
 
 public class StoreQueryResp {
-    private Integer sId;
+    private Long sId;
 
     private String sPlace;
 
@@ -13,11 +13,11 @@ public class StoreQueryResp {
 
     private String sKind;
 
-    public Integer getsId() {
+    public Long getsId() {
         return sId;
     }
 
-    public void setsId(Integer sId) {
+    public void setsId(Long sId) {
         this.sId = sId;
     }
